backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
bitarray==3.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
ckzg==2.1.5
cytoolz==1.1.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==1.0.1
eth_abi==5.2.0
frozenlist==1.8.0
idna==3.11
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lru-dict==1.2.0
multidict==6.7.0
parsimonious==0.10.0
propcache==0.4.1
protobuf==6.33.1
pycryptodome==3.23.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
pyunormalize==17.0.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
rpds-py==0.28.0
sqlparse==0.5.3
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
